fix(files): move drag and drop to sidebar tree
